# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ComicVerse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /comics/{comicId}/panels:
    get:
      operationId: listPanels
      tags: [panels]
      summary: List panels for a comic
      security:
        - clerkAuth: []
      parameters:
        - name: comicId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of panels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Panel"
    post:
      operationId: createPanel
      tags: [panels]
      summary: Create a panel
      security:
        - clerkAuth: []
      parameters:
        - name: comicId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePanelInput"
      responses:
        "201":
          description: Created panel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
  /panels/{id}:
    put:
      operationId: updatePanel
      tags: [panels]
      summary: Update a panel
      security:
        - clerkAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePanelInput"
      responses:
        "200":
          description: Updated panel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
    delete:
      operationId: deletePanel
      tags: [panels]
      summary: Delete a panel
      security:
        - clerkAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted

  # ── AI ───────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Panel:
      type: object
      properties:
        id:
          type: integer
        comicId:
          type: integer
        order:
          type: integer
        dialogue:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        imageData:
          type: string
          nullable: true
        imagePrompt:
          type: string
          nullable: true
        characterIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, comicId, order, characterIds, createdAt, updatedAt]
    CreatePanelInput:
      type: object
      properties:
        order:
          type: integer
        dialogue:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        imageData:
          type: string
          nullable: true
        imagePrompt:
          type: string
          nullable: true
        characterIds:
          type: array
          items:
            type: integer
    UpdatePanelInput:
      type: object
      properties:
        order:
          type: integer
        dialogue:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        imageData:
          type: string
          nullable: true
        imagePrompt:
          type: string
          nullable: true
        characterIds:
          type: array
          items:
            type: integer

    # ── AI ──
...